+++ src/rs/file.rs
@@ -12,8 +12,12 @@ use crate::{symbol, utils};
 #[cfg(feature = "python_bindings")]
 use pyo3::prelude::*;
 
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
 #[derive(Debug, Clone)]
 #[cfg_attr(feature = "python_bindings", pyclass(module = "mapfile_parser"))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct File {
     pub filepath: PathBuf,