+++ src/rs/symbol.rs
@@ -7,8 +7,12 @@ use std::hash::{Hash, Hasher};
 #[cfg(feature = "python_bindings")]
 use pyo3::prelude::*;
 
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
 #[derive(Debug, Clone)]
 #[cfg_attr(feature = "python_bindings", pyclass(module = "mapfile_parser"))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct Symbol {
     pub name: String,
 
@@ -20,7 +24,9 @@ pub struct Symbol {
 
     pub align: Option<u64>,
 
+    // idk if it is worth to continue maintaining this, given the complexity introduced by other features
     #[cfg(feature = "python_bindings")]
+    #[cfg(not(feature = "serde"))]
     chached_name: Option<PyObject>,
 }
 
@@ -40,6 +46,7 @@ impl Symbol {
             align,
 
             #[cfg(feature = "python_bindings")]
+            #[cfg(not(feature = "serde"))]
             chached_name: None,
         }
     }
@@ -53,6 +60,7 @@ impl Symbol {
             align: None,
 
             #[cfg(feature = "python_bindings")]
+            #[cfg(not(feature = "serde"))]
             chached_name: None,
         }
     }
@@ -144,6 +152,7 @@ pub(crate) mod python_bindings {
         /* Getters and setters */
 
         #[getter]
+        #[cfg(not(feature = "serde"))]
         fn get_name(&mut self) -> PyObject {
             Python::with_gil(|py| {
                 if self.chached_name.is_none() {
@@ -154,9 +163,18 @@ pub(crate) mod python_bindings {
             })
         }
 
+        #[getter]
+        #[cfg(feature = "serde")]
+        fn get_name(&self) -> PyResult<String> {
+            Ok(self.name.clone())
+        }
+
         #[setter]
         fn set_name(&mut self, value: String) -> PyResult<()> {
-            self.chached_name = None;
+            #[cfg(not(feature = "serde"))]
+            {
+                self.chached_name = None;
+            }
             self.name = value;
             Ok(())
         }